+++ ECSJobDemos/Assets/Packages/UnityEngine.ECS/ChunkDataUtility.cs
@@ -45,7 +45,7 @@ namespace UnityEngine.ECS
             {
                 IntPtr src = GetComponentData(srcChunk, srcIndex, i);
                 IntPtr dst = GetComponentData(dstChunk, dstIndex, i);
-                UnsafeUtility.MemCpy(dst, src, arch->sizeOfs[i]);
+                UnsafeUtility.MemCpy(dst, src, (ulong)arch->sizeOfs[i]);
             }
         }
 
@@ -67,14 +67,14 @@ namespace UnityEngine.ECS
                 for (int t = 1; t != arch->typesCount; t++)
                 {
                     IntPtr dst = GetComponentData(dstChunk, dstIndex + e, t);
-                    UnsafeUtility.MemClear(dst, arch->sizeOfs[t]);
+                    UnsafeUtility.MemClear(dst, (ulong)arch->sizeOfs[t]);
                 }
             }
         }
 
         public static void MemCpyReplicate(IntPtr dst, IntPtr src, int size, int count)
         {
-            UnsafeUtility.MemCpy(dst, src, size);
+            UnsafeUtility.MemCpy(dst, src, (ulong)size);
             if (count == 1)
                 return;
             src = dst;
@@ -84,13 +84,13 @@ namespace UnityEngine.ECS
 
             while (remainSize > copySize)
             {
-                UnsafeUtility.MemCpy(dst, src, copySize);
+                UnsafeUtility.MemCpy(dst, src, (ulong)copySize);
                 dst += copySize;
                 remainSize -= copySize;
                 copySize += copySize;
             }
 
-            UnsafeUtility.MemCpy(dst, src, remainSize);
+            UnsafeUtility.MemCpy(dst, src, (ulong)remainSize);
         }
 
         public static void ReplicateComponents(Chunk* srcChunk, int srcIndex, Chunk* dstChunk, int dstBaseIndex, int count)
@@ -133,7 +133,7 @@ namespace UnityEngine.ECS
                 {
                     IntPtr src = srcChunk->buffer + srcArch->offsets[srcI] + srcIndex * srcArch->strides[srcI];
                     IntPtr dst = dstChunk->buffer + dstArch->offsets[dstI] + dstIndex * dstArch->strides[dstI];
-                    UnsafeUtility.MemCpy(dst, src, srcArch->sizeOfs[srcI]);
+                    UnsafeUtility.MemCpy(dst, src, (ulong)srcArch->sizeOfs[srcI]);
                     ++srcI;
                     ++dstI;
                 }
